requirements.txt
jsonschema
pyyaml
wllvm
scan-build
pytest-pylint
pylint
websockets
jsonrpcserver
jsonrpcclient
pytest
